3m-usermanagement/usermanagement/requirements.txt
asgiref==3.7.2
boto3==1.34.96
botocore==1.34.96
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-cors-headers==4.3.1
django-import-export==3.3.7
django-phonenumber-field==7.3.0
django-storages==1.14.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.6
jmespath==1.0.1
MarkupPy==1.14
# mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
phonenumbers==8.13.30
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
reportlab==4.2.0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
